requirements.txt
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@stable
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
diffusers==0.31.0
transformers
accelerate
omegaconf==2.3.0
einops==0.8.1
kornia==0.8.0
opencv-python==4.11.0.86
pillow==10.4.0
scipy==1.15.2
scikit-image==0.18.3
numpy==1.24.4
matplotlib
tqdm
imageio
imageio-ffmpeg
plyfile==1.1
pexpect
lightning
utils3d==0.0.2
decord==0.6.0
fire
